\name{IDI.INF.GRAPH}
\alias{IDI.INF.GRAPH}

\title{Function to display IDI and other measures in a graph
}

\description{This function generates a plot to graphically display IDI, continuous NRI, and median improvement
}

\usage{
IDI.INF.GRAPH(x, \dots)
}

\arguments{
  \item{x}{An object generated by \code{IDI.INF}}
  \item{\dots}{Arguments passed to \code{plot()}}
}

\details{This function provide a plot to graphically display IDI, continous-NRI and median improvement.
}


\examples{

## see example in IDI.INF
  
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
