\name{IDI.INF.OUT}
\alias{IDI.INF.OUT}
\title{Function to print the summary
}
\description{This function disply a summary result performed by \code{IDI.INF()}
}

\usage{IDI.INF.OUT(x)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object generated by \code{IDI.INF}}
}
\details{This function displays the point estimates of IDI, continous-NRI and median improvement, and corresponding (1-alpha) confidence intervals.
}

%\value{}

%\references{Uno H, Tian L, Cai T, Kohane IS, and Wei LJ.
%Comparing risk scoring systems beyond the ROC paradigm in survival analysis. %Working Paper 107, Harvard University Biostatistics Working Paper Series. %http://www.bepress.com/harvardbiostat/paper107
%}

%\author{Hajime Uno
%}
%\note{}

%\seealso{\code{IDI.INF}}

\examples{

## see example in IDI.INF

}
