\name{IDI.INF.GRAPH}
\alias{IDI.INF.GRAPH}

\title{Function to display IDI and other measures in a graph
}

\description{This function generates a plot to graphically display IDI, continuous NRI, and median improvement
}

\usage{
IDI.INF.GRAPH(x, main = NULL, xlab = NULL, ylab = NULL,
cex.main = NULL, cex.lab = NULL, ...)
}

\arguments{
  \item{x}{An object generated by \code{IDI.INF}}
  \item{main}{main title of graph}
  \item{xlab}{label of x-axis. The default is \code{"s"}}
  \item{ylab}{label of y-axis. The default is
  \code{expression(paste("pr(",hat(D)<=s,")"))}
  }
  \item{cex.main}{size of the main title}
  \item{cex.lab}{size of the labels}
  \item{\dots}{Arguments passed to \code{plot()}}
}

\details{This function provide a plot to graphically display IDI, continous-NRI and median improvement.
}


\examples{

## see example in IDI.INF

}
